Prototype/WhiteBoxCapsule/backend/requirements.txt
fastapi
pydantic
python-dotenv
psycopg2-binary
sqlalchemy
uvicorn
python-jose
python-dateutil
passlib
python-multipart
